Improve translation
* One langage per file
* Move translation stuff in plugins/translation directory
